% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST21_2022.R
\name{after_setup}
\alias{after_setup}
\title{Run the MAST21 2022 protocol}
\usage{
after_setup(
  page_type = "record_midi_page",
  setup_pages = TRUE,
  data_collection_method = c("midi", "audio", "key_presses"),
  get_p_id = TRUE,
  musicassessr_aws = TRUE,
  dinosaur_instructions =
    "Please press the “record” button and read the sentence below out loud: ",
  absolute_url = "https://musicog.ca/",
  final_qualtrics_url =
    "https://upeiairs.qualtrics.com/jfe/form/SV_5vDAjJhxLqZw7Km?participant="
)
}
\arguments{
\item{final_qualtrics_url}{}
}
\description{
Run the MAST21 2022 protocol
}
