% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST21_2022.R
\name{deploy_MAST21_2022}
\alias{deploy_MAST21_2022}
\title{Run the MAST21 2022 protocol}
\usage{
deploy_MAST21_2022(
  musicassessr_state = "test",
 
    dinosaur_instructions = "Please press the “record” button and read the sentence below out loud: ",
 
    final_qualtrics_url = "https://upeiairs.qualtrics.com/jfe/form/SV_5vDAjJhxLqZw7Km?participant="
)
}
\arguments{
\item{final_qualtrics_url}{}
}
\value{

}
\description{
Run the MAST21 2022 protocol
}
