% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST21_2022_v2.R
\name{deploy_MAST21_2022_v2}
\alias{deploy_MAST21_2022_v2}
\title{Run the MAST21 2022 protocol version 2}
\usage{
deploy_MAST21_2022_v2(
  musicassessr_state = "test",
 
    dinosaur_instructions = "Please press the “record” button and read the sentence below out loud: "
)
}
\arguments{
\item{musicassessr_state}{}
}
\value{

}
\description{
Run the MAST21 2022 protocol version 2
}
