% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST21.R
\name{deploy_MAST21_wav}
\alias{deploy_MAST21_wav}
\title{deploy the MAST21 as wavs}
\usage{
deploy_MAST21_wav(musicassessr_aws = TRUE, app_name = "UPEI_MAST21")
}
\arguments{
\item{musicassessr_aws}{}
}
\description{
deploy the MAST21 as wavs
}
